requirements.txt
discord
python-dotenv

Pillow

beautifulsoup4
aiohttp
requests

sqlalchemy
aiosqlite

pytube